refactor(limit-swap): present both option sheets as sheets, and split the form by component
The custom-expiry and custom-offset sheets came up at full height over a
transparent background, showing the form through them. Neither had been
through `sheetStyle`, which is where this app's sheets get their bgPrimary
backing and their drag indicator; without it `crossPlatformSheet` hands the
content straight to a bare `.sheet` and nothing paints. Both now take
`.sheetStyle(detents: [.medium])`, and a spacer pins the Set button to the
bottom so the controls sit at the top rather than floating in the middle of
the detent. `.medium` rather than a pinned height because the notice row under
each of them is a sentence — a longer one in de and pt — and it wraps again
with Dynamic Type.

LimitSwapBodyView.swift had reached 2100 lines and twenty types. Split one
file per component, which the flat layout makes natural: each card, row, sheet
and pill is an independent piece of the same form and none of them referenced
the body view. Content moved verbatim; the only edits are dropping `private`
from the declarations that no longer share a file, and the two corner radii
moving to their own file so a card and a notice cannot drift apart now that
they are no longer neighbours. `decimalOnly()` stayed private, next to the
price card that is its only caller.
