+++ VultisigApp/VultisigApp/Features/LimitSwap/Views/LimitInlineNotice.swift
@@ -1,0 +1,37 @@
+//
+//  LimitInlineNotice.swift
+//  VultisigApp
+//
+
+import SwiftUI
+
+// MARK: - Inline notice (shared shape)
+
+struct LimitInlineNotice: View {
+
+    let systemImage: String
+    let tint: Color
+    let message: String
+
+    var body: some View {
+        HStack(spacing: 8) {
+            Image(systemName: systemImage)
+                .font(.system(size: 14))
+                .foregroundStyle(tint)
+
+            Text(message)
+                .font(Theme.fonts.caption12)
+                .foregroundStyle(Theme.colors.textSecondary)
+                .multilineTextAlignment(.leading)
+
+            Spacer(minLength: 0)
+        }
+        .padding(12)
+        .background(Theme.colors.bgSurface1)
+        .overlay(
+            limitNoticeCornerRadius.shape
+                .stroke(Theme.colors.borderLight, lineWidth: 1)
+        )
+        .clipShape(limitNoticeCornerRadius.shape)
+    }
+}
